# cleanup(linter): batch globbing to improve @nx/eslint/plugin perf